agentshield/requirements.txt
# AgentShield - AI Agent Security Platform
# Core dependencies

# MCP Server
mcp>=1.2.0

# LangGraph + LangChain
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-aws>=0.2.0
langchain-openai>=0.3.0

# Knowledge Graph
networkx>=3.2

# Data Models
pydantic>=2.9.0

# Cryptographic signing
cryptography>=43.0.0

# AWS SDK
boto3>=1.35.0

# Utilities
python-dotenv>=1.0.0

# Streamlit Dashboard
streamlit>=1.38.0
plotly>=5.24.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
